requirements.txt
streamlit==1.36.0
spotipy==2.23.0
python-dotenv==1.0.1
pandas
requests